+++ backend/canisters/group_index/api/src/updates/upload_wasm_chunk.rs
@@ -1,13 +1,12 @@
 use crate::ChildCanisterType;
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
-use types::Hash;
+use types::{CanisterWasmBytes, Hash};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
     pub canister_type: ChildCanisterType,
-    #[serde(with = "serde_bytes")]
-    pub chunk: Vec<u8>,
+    pub chunk: CanisterWasmBytes,
     pub index: u8,
 }