+++ backend/canisters/local_user_index/api/src/updates/c2c_push_wasm_chunk.rs
@@ -1,11 +1,11 @@
 use crate::ChildCanisterType;
 use serde::{Deserialize, Serialize};
+use types::CanisterWasmBytes;
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Args {
     pub canister_type: ChildCanisterType,
-    #[serde(with = "serde_bytes")]
-    pub chunk: Vec<u8>,
+    pub chunk: CanisterWasmBytes,
     pub index: u8,
 }