+++ frontend/app/src/stores/xframe.ts
@@ -24,11 +24,14 @@ const FRAME_ANCESTORS = [
     "https://spyzr-gqaaa-aaaan-qd66q-cai.icp0.io", // vaultbet
     "https://zkpoker.app", // ZKP
     "https://panoramablock.com", // Panorama Block Live
-    "https://zdgud-kqaaa-aaaal-ajn4q-cai.icp0.io", // tendy zone test
+    "https://zdgud-kqaaa-aaaal-ajn4q-cai.icp0.io", // tendyzone test
     "https://okowr-oqaaa-aaaag-qkedq-cai.icp0.io", // konecta pre-register
     "https://pre.konecta.one", //konecta pre-register
     "https://konecta.one", //konecta website/webapp
-    "https://y7mum-taaaa-aaaag-qklxq-cai.icp0.io" // konecta pre-register test env
+    "https://y7mum-taaaa-aaaag-qklxq-cai.icp0.io", // konecta pre-register test env
+    "https://dragonwizards.club", // dragon paladin wizard web2 test
+    "https://ic-vc.com", // ICVC via domain
+    "https://mnc6b-aaaaa-aaaap-qhnrq-cai.icp0.io" // ICVC via canister ID
 ];
 
 type InboundXFrameMessage = UpdateTheme | ChangeRoute | OverrideSettings | Logout;