+++ backend/canisters/local_group_index/CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+## [[2.0.1364](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1364-local_group_index)] - 2024-10-02
+
+### Changed
+
+- Add `CanisterWasmBytes` to reduce duplication ([#6480](https://github.com/open-chat-labs/open-chat/pull/6480))
+
+### Fixed
+
+- Fix chunked Community upgrades ([#6483](https://github.com/open-chat-labs/open-chat/pull/6483))
+
+## [[2.0.1363](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1363-local_group_index)] - 2024-09-30
+
 ### Added
 
 - Support upgrading to large wasms by uploading in chunks ([#6453](https://github.com/open-chat-labs/open-chat/pull/6453))