+++ backend/canisters/local_user_index/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Add `CanisterWasmBytes` to reduce duplication ([#6480](https://github.com/open-chat-labs/open-chat/pull/6480))
+
+## [[2.0.1362](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1362-local_user_index)] - 2024-09-30
+
 ### Added
 
 - Support upgrading to large wasms by uploading in chunks ([#6453](https://github.com/open-chat-labs/open-chat/pull/6453))